Node - Add utils::slice
